requirements.txt
requests >=2.31.0
jsonschema >=4.21.1
allure-pytest >= 2.13.5
pytest >= 8.3.2
selene >= 2.0.0rc9
python-dotenv ~= 1.0.1